+++ utilities/inc/dev/inc/browser-sync.php
@@ -13,17 +13,25 @@ add_action( 'wp_footer', function() {
     // check if BrowserSync is actually running to avoid browser timeout
     // @since 1.20.3
     foreach ( [
-        explode( ':', $_SERVER['HTTP_HOST'] )[0],
-        'host.docker.internal',
-        '172.17.0.1',
-        'localhost'
+            explode( ':', $_SERVER['HTTP_HOST'] )[0],
+            'host.docker.internal',
+            '172.17.0.1',
+            'localhost'
     ] as $host ) {
         if ( $socket = @fsockopen( $host, 3000, $errno, $errstr, .1 ) ) { ?>
             <script id="__bs_script__">//<![CDATA[
                 document.write( '<script async src="http://' + location.hostname + ':3000/browser-sync/browser-sync-client.js?v=2.27.10"><\/script>' );
-            //]]></script>
+                //]]></script>
 
-            <?php break;
+            <?php echo minify_css( "<style>
+                #__bs_notify__ {
+                    font-size: 12px !important;
+                    padding: .5em 1em !important;
+                    opacity: .65;
+                }
+            </style>" );
+
+            break;
         }
     }